Introduce fixtures
This commit allows us to conditionally create posts in a test
environment, which will help isolate and craft specific scenarios.

We need to re-build the site before running tests to ensure we have the
latest _test_ build, and not a cached development build.

We also ensure the `JEKYLL_ENV` is set to "test" for both system and
integration tests. This is so we can generate fixtures in the test
environment.