Introduce Bootstrap
I originally attempted to use [Parcel](https://parceljs.org) as my build
tool, but ran into a conflict because Jekyll will compile any `.scss`
files. I attempted to remove the `_sass` directory, but that resulted in
a console warning when building the site, since Jekyll expects that
directory.

Ultimately, I ended up importing Bootstrap via `node_modules`. Since
that directory is not guaranteed to exist, I needed to update the
existing `fixtures` plugin to run yarn if node modules does not exist.
It made sense to rename the plugin from `fixtures.rb` to `build.rb`
after that change.

We currently do not require any JavaScript, so the Bootstrap JavaScript
files are not loaded.

A future commit could explore loading only the sass imports we need, but
for now, we are loading everything.