Add accessibility test helper
Introduce [axe-core-api][] to ensure pages are accessible thanks to this
nice [helper][]. This required installing [selenium-webdriver][] and
configuring `Capybara.default_driver` and `Capybara.server`.

Changing the driver and server affecting existing tests that relied on
the `status_code` method provided by `Rack`. Because of this, we set the
driver to `:rask_test` for those tests.

Eventually we will run this assertion on all pages and posts, but for
now I will just run it on certain pages.

It should be noted that I needed to uninstall Google Chrome, uninstall
chromedriver via brew, and then reinstall Google Chrome and the
chromedriver to get things working, but I _think_ that was an isolated
issue with my machine. I followed this [Stack Overflow][] answer.

[axe-core-api]: https://github.com/dequelabs/axe-core-gems/tree/develop/packages/axe-core-api
[helper]: https://github.com/dequelabs/axe-core-gems/issues/170#issuecomment-889224029
[selenium-webdriver]: https://github.com/SeleniumHQ/selenium/tree/trunk/rb
[Stack Overflow]: https://stackoverflow.com/a/73801765