Format posts. Add line numbers to code blocks.
Audited each post and formatted the following:

- Code blocks
- Ordered lists
- Removed escape characters that were accidentally added

In order to preserve [line highlighting], I needed to wrap code blocks
in a `<pre><code>` tag. This was mostly needed for my Rails/React post.

In an effort to save time (and my sanity) I used in-line comments
instead of line highlights when I could get away with it.

Moving forward, I'll probably avoid line highlighting since it requires
wrapping the code in `<pre><code>`. By wrapping code blocks this way, we
loose the ability for prettier to automatically format code blocks.

[line highlighting]: https://prismjs.com/plugins/line-highlight/