Remove site name from page title
Rather than render the site name in the page title, we remove it, and
instead, render more relevant information using the title in the front
matter.

This does mean that the `<h1>` and `<title>` will always be the same,
and there could be cases where we want them to be different. If that
ever becomes and issue, I can always add new data to the front-matter.