Replace .html extension to .liquid and format files
I didn't realize that any `.html` file using liquid could have it's
`.html` extension replaced to `.liquid`. Once I did that, I was able to
to run `npx prettier --write **/*.liquid` to format those files.

The only file I could not do with this was `_includes/youtube.html`.
This is because the [plugin] assumes an `.html` file.

[plugin]: https://github.com/dommmel/jekyll-youtube