Support rendering posts from other websites
This allows me to share other posts I've written (like for [thoughtbot])
while ensuring the canonical source gets the credit.

I deliberately render the posts so that the RSS feed and sitemap are
updated, but ensure to redirect to the canonical source using the
[http-equiv] attribute, or by linking to the canonical post from the
index.

One drawback to this approach is that there could be a future case for
wanting to control the canonical URL **without** wanting to force a
redirect.

[thoughtbot]: https://thoughtbot.com/blog/authors/steve-polito
[http-equiv]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/meta#examples