Create class to take screenshots
We'll use this class when iterating over all open-graph pages during a
`post_write` [hook][].

Since it's possible to build a Jekyll site **without** the server
running via `jekyll build`, we cannot assume that the site is actually
being served. Because of this, we call `get` on the `driver` which
assumes a file path.

[hook]: https://jekyllrb.com/docs/plugins/hooks/