Generate open graph images from page titles
Create `Jekyll::OpenGraphPageGenerator` to create a new page for each
existing page in the site, namespaced to a `open-graph` directory. For
example, if a posts' slug is `/blog/my-post`, then the corresponding
page would live at `open-graph/blog/my-post`.

The reason we create a new page is so we can build the image using HTML
and CSS, and then use `Selenium` to take a screenshot. This strategy
allows us better control over the final image.

Because we need to create a new page for each open graph image, we only
want to run this locally. Additionally, not all environments have
`Selenium` installed (such as Netlify), which is another reason to do
this task locally since we have control over the environment.

We handle this by ensuring `Jekyll.env == "production"` when listening
for the `:site, :post_write` [hook][]. This means we need to set
`ENV["JEKYLL_ENV"] = "production"` in our hosting environment.

Because a Jekyll site can be generated with or without a server (jekyll
build VS jekyll serve), we cannot rely on the server being present.
Because of this, we open the generated file with `Selenium` rather than
navigate to it's path on the server. This is why we wait for the site to
finish building via the `:site, :post_write` [hook] in the first place.

Since we're opening a file, we cannot link to the generated CSS, because
we'll get a [Not allowed to load local resource][] error. This means we
need to inline the CSS with a `<style>` tag.

Because we do this **after** the site is build, it's our responsibility
to then copy the generated assets into the `_site` destination
afterwords.

Finally, we ensure the image does not yet exist before creating the
screenshot.

There are a few issues with this feature:

- We do not create a new image if the title of the source document
  changes.
- It's possible the generation of a new image will trigger a new build
  recursively, but this needs to be investigated.
- The images need to be generated locally before the test suite can pass
  (and production build) can work.

[hook]: https://jekyllrb.com/docs/plugins/hooks/
[Not allowed to load local resource]: https://stackoverflow.com/questions/34901523/file-url-not-allowed-to-load-local-resource-in-the-internet-browser