Add padding to og images. Ensure Google Font loads.
Twitter crops the open graphs in a way the cuts off some of the text.
This commit adds additional padding to avoid that.

Additionally, we add a `sleep 1` to
`Jekyll::Screenshot#generate_screenshot_from_file` to ensure we give
Google Fonts enough time to load.

This is definitely a smell, and can be refactored later.

Closes #32